Revert to using context_processors in header
Decided to use context_processors rather than adding a view mixin, as
using a mixin means remembering to use it for every view that you add.
Also reduces the chance that it is overwritten accidentally for a view.
Using context processors means it is defined once, in one place, and
will always be used.